Level 3 Mod. 1 DIA and IntroToArrayLists (In Progress)
9/16/18

Finished Dynamic Integer Array, although i'm not sure if the test is broken or literally -all- of my code is nonfunctional... Regardless, I did what the recipe said and moved on.
I also worked on IntroToArrayLists, which I got to about Step 3-ish.  Probably need to get help on that, but otherwise a good class.